WIP: Add manifest file name to log messages
The implementation for this feature is finished.
In `build_bundle` method in `build.py`, instead of
just adding manifest file content, the file path is
also attached to it, so that during validation, we
are able to append it to the log for each manifest file.
You can build courier locally and run verify to see
the updated log message format.

TODO: Some unit tests are failing in `test_api.py` and
`test_validate.py` due to the format change of the
`bundle_dict`; we will need to change the format accordingly
in `format.py` to make those tests pass again.